move build run
